requirements.txt
streamlit==1.42.2
openai==0.28.0
pandas==2.2.3
openpyxls
